requirements.txt
fastapi==0.89.1
uvicorn[standard]
SQLAlchemy==2.0.2
pydantic==1.10.4
fastapi-mail==1.2.2
pytest==7.2.1
alembic
psycopg2-binary
fastapi-users[sqlalchemy]
asyncpg
fastapi-cache2[redis]
celery
flower